move siacoin errors to dedicated siacoin::error module
refactors various errors to fit new error handling patterns
